# Refactor `toNormalisedLower`: shorter and slightly faster.