This is a follow up on https://github.com/prometheus/prometheus/pull/14170

TL;DR, this version of code is much shorter, easier to understand (IMO) and it's also faster!

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/labels
                                                             │     main      │                 new                 │
                                                             │    sec/op     │   sec/op     vs base                │
ToNormalizedLower/length=10/uppercase=none/ascii=true-12       19.840n ±  0%   5.546n ± 0%  -72.05% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=none/ascii=false-12       138.2n ±  0%   120.7n ± 1%  -12.63% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=first/ascii=true-12       20.66n ±  1%   18.52n ± 5%  -10.36% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=first/ascii=false-12      142.6n ±  1%   145.0n ± 1%   +1.72% (p=0.004 n=10)
ToNormalizedLower/length=10/uppercase=last/ascii=true-12        21.60n ±  0%   19.54n ± 1%   -9.52% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=last/ascii=false-12       150.3n ±  4%   132.6n ± 1%  -11.81% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=all/ascii=true-12         26.14n ±  2%   20.84n ± 1%  -20.31% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=all/ascii=false-12        143.8n ±  1%   130.8n ± 1%   -9.01% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=none/ascii=true-12       67.97n ±  0%   48.21n ± 1%  -29.08% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=none/ascii=false-12      1.743µ ±  2%   1.719µ ± 1%   -1.38% (p=0.017 n=10)
ToNormalizedLower/length=100/uppercase=first/ascii=true-12      69.95n ±  2%   67.11n ± 1%   -4.07% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=first/ascii=false-12     1.721µ ±  0%   1.797µ ± 0%   +4.42% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=last/ascii=true-12       71.70n ± 14%   68.12n ± 1%   -4.99% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=last/ascii=false-12      1.779µ ±  2%   1.728µ ± 1%   -2.87% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=all/ascii=true-12       147.70n ±  1%   99.92n ± 0%  -32.35% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=all/ascii=false-12       1.632µ ±  2%   1.652µ ± 0%   +1.19% (p=0.022 n=10)
ToNormalizedLower/length=1000/uppercase=none/ascii=true-12      550.6n ±  4%   407.6n ± 0%  -25.96% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=none/ascii=false-12     14.79µ ±  1%   14.50µ ± 1%   -1.95% (p=0.002 n=10)
ToNormalizedLower/length=1000/uppercase=first/ascii=true-12     554.6n ±  1%   554.8n ± 1%        ~ (p=0.927 n=10)
ToNormalizedLower/length=1000/uppercase=first/ascii=false-12    14.67µ ±  0%   15.14µ ± 0%   +3.21% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=last/ascii=true-12      542.8n ±  2%   551.5n ± 0%   +1.60% (p=0.027 n=10)
ToNormalizedLower/length=1000/uppercase=last/ascii=false-12     14.74µ ±  0%   14.55µ ± 0%   -1.30% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=all/ascii=true-12      1205.5n ±  1%   814.5n ± 1%  -32.43% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=all/ascii=false-12      13.76µ ±  1%   13.66µ ± 0%   -0.72% (p=0.035 n=10)
ToNormalizedLower/length=4000/uppercase=none/ascii=true-12      2.173µ ±  0%   1.606µ ± 0%  -26.09% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=none/ascii=false-12     58.72µ ±  1%   58.03µ ± 2%   -1.17% (p=0.035 n=10)
ToNormalizedLower/length=4000/uppercase=first/ascii=true-12     2.196µ ±  1%   2.177µ ± 0%   -0.87% (p=0.001 n=10)
ToNormalizedLower/length=4000/uppercase=first/ascii=false-12    58.27µ ±  0%   59.97µ ± 0%   +2.93% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=last/ascii=true-12      2.147µ ±  1%   2.095µ ± 5%        ~ (p=0.102 n=10)
ToNormalizedLower/length=4000/uppercase=last/ascii=false-12     58.73µ ±  1%   58.20µ ± 1%   -0.90% (p=0.001 n=10)
ToNormalizedLower/length=4000/uppercase=all/ascii=true-12       4.800µ ±  0%   3.265µ ± 1%  -31.97% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=all/ascii=false-12      54.47µ ±  0%   54.53µ ± 1%        ~ (p=0.853 n=10)
geomean                                                         951.6n         832.9n       -12.47%

                                                             │     main     │                    new                    │
                                                             │     B/op     │     B/op      vs base                     │
ToNormalizedLower/length=10/uppercase=none/ascii=true-12         16.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=none/ascii=false-12       25.000 ± 0%     9.000 ± 0%   -64.00% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=first/ascii=true-12        16.00 ± 0%     16.00 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=10/uppercase=first/ascii=false-12       20.00 ± 0%     17.00 ± 0%   -15.00% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=last/ascii=true-12         16.00 ± 0%     16.00 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=10/uppercase=last/ascii=false-12        27.00 ± 0%     11.00 ± 0%   -59.26% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=all/ascii=true-12          16.00 ± 0%     16.00 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=10/uppercase=all/ascii=false-12         32.00 ± 0%     22.00 ± 0%   -31.25% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=none/ascii=true-12        112.0 ± 0%       0.0 ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=none/ascii=false-12     1.159Ki ± 0%   1.050Ki ± 0%    -9.44% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=first/ascii=true-12       112.0 ± 0%     112.0 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=100/uppercase=first/ascii=false-12    1.126Ki ± 0%   1.104Ki ± 0%    -1.91% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=last/ascii=true-12        112.0 ± 0%     112.0 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=100/uppercase=last/ascii=false-12     1.170Ki ± 0%   1.061Ki ± 0%    -9.35% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=all/ascii=true-12         112.0 ± 0%     112.0 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=100/uppercase=all/ascii=false-12      1.203Ki ± 0%   1.137Ki ± 0%    -5.52% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=none/ascii=true-12     1.000Ki ± 0%   0.000Ki ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=none/ascii=false-12    5.862Ki ± 0%   4.862Ki ± 0%   -17.06% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=first/ascii=true-12    1.000Ki ± 0%   1.000Ki ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=1000/uppercase=first/ascii=false-12   5.524Ki ± 0%   5.375Ki ± 0%    -2.70% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=last/ascii=true-12     1.000Ki ± 0%   1.000Ki ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=1000/uppercase=last/ascii=false-12    5.975Ki ± 0%   4.975Ki ± 0%   -16.74% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=all/ascii=true-12      1.000Ki ± 0%   1.000Ki ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=1000/uppercase=all/ascii=false-12     6.312Ki ± 0%   5.712Ki ± 0%    -9.51% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=none/ascii=true-12     4.000Ki ± 0%   0.000Ki ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=none/ascii=false-12    21.41Ki ± 0%   17.41Ki ± 0%   -18.68% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=first/ascii=true-12    4.000Ki ± 0%   4.000Ki ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=4000/uppercase=first/ascii=false-12   19.99Ki ± 0%   19.49Ki ± 0%    -2.50% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=last/ascii=true-12     4.000Ki ± 0%   4.000Ki ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=4000/uppercase=last/ascii=false-12    21.89Ki ± 0%   17.89Ki ± 0%   -18.28% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=all/ascii=true-12      4.000Ki ± 0%   4.000Ki ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=4000/uppercase=all/ascii=false-12     23.31Ki ± 0%   20.91Ki ± 0%   -10.30% (p=0.000 n=10)
geomean                                                          647.2                      ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

                                                             │    main    │                   new                   │
                                                             │ allocs/op  │ allocs/op   vs base                     │
ToNormalizedLower/length=10/uppercase=none/ascii=true-12       1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=none/ascii=false-12      1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=first/ascii=true-12      1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=10/uppercase=first/ascii=false-12     1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=10/uppercase=last/ascii=true-12       1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=10/uppercase=last/ascii=false-12      1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=all/ascii=true-12        1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=10/uppercase=all/ascii=false-12       2.000 ± 0%   1.000 ± 0%   -50.00% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=none/ascii=true-12      1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=none/ascii=false-12     5.000 ± 0%   4.000 ± 0%   -20.00% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=first/ascii=true-12     1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=100/uppercase=first/ascii=false-12    5.000 ± 0%   5.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=100/uppercase=last/ascii=true-12      1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=100/uppercase=last/ascii=false-12     5.000 ± 0%   4.000 ± 0%   -20.00% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=all/ascii=true-12       1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=100/uppercase=all/ascii=false-12      6.000 ± 0%   5.000 ± 0%   -16.67% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=none/ascii=true-12     1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=none/ascii=false-12    5.000 ± 0%   4.000 ± 0%   -20.00% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=first/ascii=true-12    1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=1000/uppercase=first/ascii=false-12   5.000 ± 0%   5.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=1000/uppercase=last/ascii=true-12     1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=1000/uppercase=last/ascii=false-12    5.000 ± 0%   4.000 ± 0%   -20.00% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=all/ascii=true-12      1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=1000/uppercase=all/ascii=false-12     6.000 ± 0%   5.000 ± 0%   -16.67% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=none/ascii=true-12     1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=none/ascii=false-12    5.000 ± 0%   4.000 ± 0%   -20.00% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=first/ascii=true-12    1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=4000/uppercase=first/ascii=false-12   5.000 ± 0%   5.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=4000/uppercase=last/ascii=true-12     1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=4000/uppercase=last/ascii=false-12    5.000 ± 0%   4.000 ± 0%   -20.00% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=all/ascii=true-12      1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=4000/uppercase=all/ascii=false-12     6.000 ± 0%   5.000 ± 0%   -16.67% (p=0.000 n=10)
geomean                                                        1.901                    ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean
```

<details> 

<summary>Old description, in case you want to understand why there are so many commits here</summary>
In the series of "optimising methods nobody cares about", I spent some optimising this one. I started from removing the redundant `isASCII` check, but I felt I couldn't just sent a PR on that so I tried to make it shorter and hopefully faster. 

I tried several approaches, some with trade-off, finally ended in a version that optimizes the performance for an already-lowercase string, while not penalizing the "uppercase" in more than 20%.

<details>
<summary>Expand to see the benchmarks results</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/labels
                                                             │     main      │                 new                 │
                                                             │    sec/op     │   sec/op     vs base                │
ToNormalizedLower/length=10/uppercase=none/ascii=true-12       19.840n ±  0%   5.356n ± 2%  -73.00% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=none/ascii=false-12       138.2n ±  0%   119.0n ± 1%  -13.90% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=first/ascii=true-12       20.66n ±  1%   20.48n ± 1%   -0.82% (p=0.019 n=10)
ToNormalizedLower/length=10/uppercase=first/ascii=false-12      142.6n ±  1%   140.2n ± 1%   -1.61% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=last/ascii=true-12        21.60n ±  0%   22.35n ± 0%   +3.50% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=last/ascii=false-12       150.3n ±  4%   129.3n ± 1%  -13.94% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=all/ascii=true-12         26.14n ±  2%   29.07n ± 1%  +11.21% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=all/ascii=false-12        143.8n ±  1%   129.8n ± 1%   -9.74% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=none/ascii=true-12       67.97n ±  0%   39.77n ± 3%  -41.48% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=none/ascii=false-12      1.743µ ±  2%   1.671µ ± 6%   -4.10% (p=0.014 n=10)
ToNormalizedLower/length=100/uppercase=first/ascii=true-12      69.95n ±  2%   68.27n ± 2%   -2.41% (p=0.009 n=10)
ToNormalizedLower/length=100/uppercase=first/ascii=false-12     1.721µ ±  0%   1.756µ ± 0%   +2.06% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=last/ascii=true-12       71.70n ± 14%   70.92n ± 0%        ~ (p=0.122 n=10)
ToNormalizedLower/length=100/uppercase=last/ascii=false-12      1.779µ ±  2%   1.700µ ± 1%   -4.47% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=all/ascii=true-12        147.7n ±  1%   163.5n ± 2%  +10.73% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=all/ascii=false-12       1.632µ ±  2%   1.627µ ± 0%   -0.31% (p=0.002 n=10)
ToNormalizedLower/length=1000/uppercase=none/ascii=true-12      550.6n ±  4%   401.5n ± 2%  -27.07% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=none/ascii=false-12     14.79µ ±  1%   14.26µ ± 0%   -3.58% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=first/ascii=true-12     554.6n ±  1%   548.4n ± 0%   -1.13% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=first/ascii=false-12    14.67µ ±  0%   14.91µ ± 1%   +1.68% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=last/ascii=true-12      542.8n ±  2%   556.1n ± 4%        ~ (p=0.796 n=10)
ToNormalizedLower/length=1000/uppercase=last/ascii=false-12     14.74µ ±  0%   14.44µ ± 0%   -2.01% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=all/ascii=true-12       1.206µ ±  1%   1.445µ ± 1%  +19.87% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=all/ascii=false-12      13.76µ ±  1%   13.60µ ± 0%   -1.15% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=none/ascii=true-12      2.173µ ±  0%   1.594µ ± 1%  -26.67% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=none/ascii=false-12     58.72µ ±  1%   56.95µ ± 1%   -3.01% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=first/ascii=true-12     2.196µ ±  1%   2.159µ ± 1%   -1.71% (p=0.002 n=10)
ToNormalizedLower/length=4000/uppercase=first/ascii=false-12    58.27µ ±  0%   59.81µ ± 0%   +2.64% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=last/ascii=true-12      2.147µ ±  1%   2.198µ ± 3%        ~ (p=0.617 n=10)
ToNormalizedLower/length=4000/uppercase=last/ascii=false-12     58.73µ ±  1%   57.30µ ± 4%   -2.43% (p=0.022 n=10)
ToNormalizedLower/length=4000/uppercase=all/ascii=true-12       4.800µ ±  0%   5.691µ ± 1%  +18.57% (p=0.000 n=10)
ToNormalizedLower/length=4000/uppercase=all/ascii=false-12      54.47µ ±  0%   54.02µ ± 0%   -0.82% (p=0.000 n=10)
geomean                                                         951.6n         880.6n        -7.46%

                                                             │     main     │                    new                    │
                                                             │     B/op     │     B/op      vs base                     │
ToNormalizedLower/length=10/uppercase=none/ascii=true-12         16.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=none/ascii=false-12       25.000 ± 0%     9.000 ± 0%   -64.00% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=first/ascii=true-12        16.00 ± 0%     16.00 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=10/uppercase=first/ascii=false-12       20.00 ± 0%     17.00 ± 0%   -15.00% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=last/ascii=true-12         16.00 ± 0%     16.00 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=10/uppercase=last/ascii=false-12        27.00 ± 0%     11.00 ± 0%   -59.26% (p=0.000 n=10)
ToNormalizedLower/length=10/uppercase=all/ascii=true-12          16.00 ± 0%     16.00 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=10/uppercase=all/ascii=false-12         32.00 ± 0%     22.00 ± 0%   -31.25% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=none/ascii=true-12        112.0 ± 0%       0.0 ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=none/ascii=false-12     1.159Ki ± 0%   1.050Ki ± 0%    -9.44% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=first/ascii=true-12       112.0 ± 0%     112.0 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=100/uppercase=first/ascii=false-12    1.126Ki ± 0%   1.104Ki ± 0%    -1.91% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=last/ascii=true-12        112.0 ± 0%     112.0 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=100/uppercase=last/ascii=false-12     1.170Ki ± 0%   1.061Ki ± 0%    -9.35% (p=0.000 n=10)
ToNormalizedLower/length=100/uppercase=all/ascii=true-12         112.0 ± 0%     112.0 ± 0%         ~ (p=1.000 n=10) ¹
ToNormalizedLower/length=100/uppercase=all/ascii=false-12      1.203Ki ± 0%   1.137Ki ± 0%    -5.52% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=none/ascii=true-12     1.000Ki ± 0%   0.000Ki ± 0%  -100.00% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=none/ascii=false-12    5.862Ki ± 0%   4.862Ki ± 0%   -17.06% (p=0.000 n=10)
ToNormalizedLower/length=1000/uppercase=first/ascii=true-12    1.000Ki ± 0%   1.000Ki ± 0%         ~ (p=1.000 n=10) ¹
…