@colega 2024-06-18
> Wondering how different the performance would be if we used strings.Builder instead.

`strings.Builder` was in one of the intermediate commits, we can't modify string.Builder's contents (boo!), so we need to write the string as we go, instead of this approach where we can just copy the entire string (and rely on CPU & compiler optimizations to do that efficiently) and then modify the uppercase chars.